avito 0.047... possible

+++ avito/AvitoProject/src/main/scala/com/sparkydots/kaggle/avito/functions/UdfFunctions.scala
@@ -18,7 +18,7 @@ object UdfFunctions extends Serializable {
 
   val parseParams = udf[Seq[Int], String](_parseParams)
 
-  val error = udf[Double, Double, Int](_error)
+  val error = udf[Double, Double, Double](_error)
 
 }
